requirements.txt
numpy>=1.26.3
scipy>=1.12
pandas>=2.1.4
matplotlib>=3.8.2
loguru>=0.7.2
pyyaml>=6.0.1